refactor(agent): unify memory under agent directory
