+++ server/cmd/server/memory_wakeup_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/1024XEngineer/XE3-ESL/server/internal/memory"
+	"github.com/1024XEngineer/XE3-ESL/server/internal/agent/memory"
 )
 
 func TestMemoryWorkerWakeupCoalescesWithoutBlocking(t *testing.T) {
