+++ server/internal/agent/memory/extraction_barrier.go
