+++ server/internal/agent/memory/extraction_barrier_test.go
