+++ server/internal/agent/memory/extraction_worker.go
