+++ server/internal/agent/memory/extractor_test.go
