+++ server/internal/agent/memory/index.go
