+++ server/internal/agent/memory/index_worker.go
