+++ server/internal/agent/memory/index_worker_test.go
