+++ server/internal/agent/memory/model.go
