+++ server/internal/agent/memory/model_test.go
