+++ server/internal/agent/memory/search.go
