+++ server/internal/agent/memory/search_test.go
