+++ server/internal/agent/memory/stable_profile.go
