+++ server/internal/agent/memory/stable_profile_test.go
