+++ server/internal/bootstrap/agent_data.go
@@ -10,6 +10,7 @@ import (
 	agentapp "github.com/1024XEngineer/XE3-ESL/server/internal/agent/app"
 	"github.com/1024XEngineer/XE3-ESL/server/internal/agent/core"
 	agentimage "github.com/1024XEngineer/XE3-ESL/server/internal/agent/image"
+	"github.com/1024XEngineer/XE3-ESL/server/internal/agent/memory"
 	agentpersistence "github.com/1024XEngineer/XE3-ESL/server/internal/agent/persistence"
 	agentruntime "github.com/1024XEngineer/XE3-ESL/server/internal/agent/runtime"
 	agentsummary "github.com/1024XEngineer/XE3-ESL/server/internal/agent/summary"
@@ -23,7 +24,6 @@ import (
 	"github.com/1024XEngineer/XE3-ESL/server/internal/identity"
 	"github.com/1024XEngineer/XE3-ESL/server/internal/matter"
 	matteragenttool "github.com/1024XEngineer/XE3-ESL/server/internal/matter/agenttool"
-	"github.com/1024XEngineer/XE3-ESL/server/internal/memory"
 	"github.com/1024XEngineer/XE3-ESL/server/internal/review"
 	reviewagenttool "github.com/1024XEngineer/XE3-ESL/server/internal/review/agenttool"
 	"github.com/jackc/pgx/v5/pgxpool"
