+++ server/internal/bootstrap/memory_context.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 
+	"github.com/1024XEngineer/XE3-ESL/server/internal/agent/memory"
 	agentruntime "github.com/1024XEngineer/XE3-ESL/server/internal/agent/runtime"
-	"github.com/1024XEngineer/XE3-ESL/server/internal/memory"
 )
 
 type agentMemoryContextSearcher struct {
