+++ server/internal/bootstrap/memory_context_test.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/1024XEngineer/XE3-ESL/server/internal/agent/memory"
 	agentruntime "github.com/1024XEngineer/XE3-ESL/server/internal/agent/runtime"
-	"github.com/1024XEngineer/XE3-ESL/server/internal/memory"
 	"github.com/1024XEngineer/XE3-ESL/server/internal/platform/requestcontext"
 )
